test(compiler-cli): improve compliance test performance (#39956)
The newly built compliance test runner was not using the shared source
file cache that was added in b627f7f02e714f68a2f6ab0724597ad6107db3ce,
which offers a significant performance boost to the compliance test
targets.

PR Close #39956
